Solve Quest 2 Part 2
We reversed all the words and added them to the original words to handle the "both directions" part of the problem.

I just used a `HashSet` to keep track of all the positions in the text that were matched by some word to handle the counting. That may not have been terribly efficient, but it works fine for now.

Updating the parsing to work with multi-line text was super easy using `splitn`.